put showdb into getpathy then called getpathy instead of showdb in database.js.
should mean it has to add a value to hopeitwork before it starts the
showdb process. currently, getpathy in downloader gets a value for
hopeitwork before showdb launches (maybe just faster? so no guarantee
it'll always manage it), but nothing in database.js that finds a value,
except for in querysuccess (which also seems to cause problems with it
appearing after hit exit)